+++ extensions/Worker.Extensions.Storage.Blobs/release_notes.md
@@ -2,3 +2,4 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
+- Overriding binding property name for "BlobPath" property of BlobTrigger (#1086)